requirements.txt
# Core dependencies
python-dotenv==1.0.0
langchain
langgraph==0.0.26
google-generativeai==0.3.2

# GitHub API
PyGithub==2.1.1
requests==2.31.0

# Slack bot
slack-bolt==1.18.0
slack-sdk==3.26.1

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# Data processing
pandas==2.1.4
numpy==1.26.2

# Visualization
matplotlib==3.8.2
plotly==5.18.0
kaleido==0.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
